+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Added
-- Added `load!` function to return raw value without `:ok/:error` tuple.
 - Wrapped `Application` functions for fetching config (`get_env/3`, `fetch_env/2` and `fetch_env!/2`).
+- Added `load!` function to return raw value without `:ok/:error` tuple.
 
 ## [0.2.0] - 2016-11-21
 ### Added