+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.3.0] - 2016-11-22
 ### Added
 - Wrapped `Application` functions for fetching config (`get_env/3`, `fetch_env/2` and `fetch_env!/2`).
 - Added `load!` function to return raw value without `:ok/:error` tuple.